fear: Add Mairui service and stock search feature
In this commit, the Mairui service is added to the application, and a new feature for searching the stock is introduced. As a result, the StockStrategyController now has an async function to search the stock. A corresponding endpoint has also been added to the server's routes. The license for the Mairui API is stored in the config/index file.